Little better test
The initialization order is undetermined, and is difficult to
reproduce in a test, but this one has a better chance of hitting
the failure situation where the virtual_memory_page_size variable
is accessed before it has been set, yielding a value of zero.

This is basically the crash scenario that I encountered, which
prompted these changes.