Add force_unlock() to no_mutex.
allocator_storage takes a Mutex type supports force_unlock() but when
the allocator is stateless, the actual mutex type will be no_mutex
instead of the supported one.

We could SFINAE on teh passed-in mutex type, and then detect the
no_mutex type and turn it into a no-op. However, it doesn't seem
like a bad thing to just add force_unlock() to the no_mutex types,
since there is no locking going on anyway.